+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## New features
 
-- `find_remotes()` and `set_remotes_to_desc()` now detects "r-universe" repositories.
+- `find_remotes()` now informs when using "r-universe" repositories.
 
 # attachment 0.3.0